refactor : 상대시간 구하는 함수 이동
getRelativeTime 함수가 여러군데에서 쓰이기때문에 lib파일에 정리
